add mvp and tests

+++ src/main/java/com/example/restaurantbookingservice/service/BookingService.java
@@ -1,8 +1,13 @@
 package com.example.restaurantbookingservice.service;
 
 import com.example.restaurantbookingservice.model.Booking;
+import com.example.restaurantbookingservice.model.User;
 import com.example.restaurantbookingservice.repository.BookingRepository;
+import com.example.restaurantbookingservice.repository.UserRepository;
+import com.example.restaurantbookingservice.security.services.UserDetailsImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,12 +18,22 @@ public class BookingService {
     @Autowired
     private BookingRepository bookingRepository;
 
+    @Autowired
+    private UserRepository userRepository;
+
     public List<Booking> getAllBookings() {
         return bookingRepository.findAll();
     }
 
     public Booking getBookingById(Long id) {
-        return bookingRepository.findById(id).orElse(null);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
+
+        if (userDetails.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"))) {
+            return bookingRepository.findById(id).orElse(null);
+        } else {
+            return bookingRepository.findByUserIdAndId(userDetails.getId(), id).orElse(null);
+        }
     }
 
     public List<Booking> getBookingsByTimeSlotId(Long timeSlotId) {
@@ -26,14 +41,26 @@ public class BookingService {
     }
 
     public Booking addBooking(Booking booking) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
+        User user = userRepository.findById(userDetails.getId()).orElseThrow(() -> new RuntimeException("User not found"));
+
         List<Booking> bookings = bookingRepository.findByTimeSlotId(booking.getTimeSlot().getId());
         if (bookings.isEmpty()) {
+            booking.setUser(user);
             return bookingRepository.save(booking);
         }
         return null;
     }
 
     public void deleteBooking(Long id) {
-        bookingRepository.deleteById(id);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
+
+        if (userDetails.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"))) {
+            bookingRepository.deleteById(id);
+        } else {
+            bookingRepository.findByUserIdAndId(userDetails.getId(), id).ifPresent(bookingRepository::delete);
+        }
     }
 }
